moved admin controller methods and admin services into corresponding product and category services, checked api's in postman
