+++ ranchsupply-backend/ranchsupply-backend/src/main/java/org/launchcode/ranchsupply/controller/CategoryController.java
@@ -24,9 +24,39 @@ public class CategoryController {
         return ResponseEntity.ok(categories);
     }
 
+    @GetMapping("/{categoryId}")
+    public ResponseEntity<CategoryDto> getCategoryById(@PathVariable Long categoryId) {
+        CategoryDto categoryDto = categoryService.getCategoryById(categoryId);
+        return ResponseEntity.ok(categoryDto);
+    }
+
     @PostMapping
     public ResponseEntity<CategoryDto> addCategory(@Valid @RequestBody CategoryDto categoryDto){
         CategoryDto savedCategory = categoryService.addCategory(categoryDto);
         return new ResponseEntity<>(savedCategory, HttpStatus.CREATED);
+    }
+
+    @PutMapping("/{categoryId}")
+    public ResponseEntity<CategoryDto> updateCategory(@PathVariable Long categoryId, @RequestBody CategoryDto categoryDto) {
+        System.out.println("Update Category endpoint hit with categoryId: " + categoryId);
+        try {
+            CategoryDto updatedCategory = categoryService.updateCategory(categoryId, categoryDto);
+            return ResponseEntity.ok(updatedCategory);
+        } catch (Exception e) {
+            System.out.println("Error updating category: " + e.getMessage());
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+    }
+
+    @DeleteMapping("/{categoryId}")
+    public ResponseEntity<Void> deleteCategory(@PathVariable Long categoryId) {
+        System.out.println("Delete Category endpoint hit with categoryId: " + categoryId);
+        try {
+            categoryService.deleteCategory(categoryId);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        } catch (Exception e) {
+            System.out.println("Error deleting category: " + e.getMessage());
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 }
