+++ ranchsupply-backend/ranchsupply-backend/src/main/java/org/launchcode/ranchsupply/controller/ProductController.java
@@ -22,7 +22,6 @@ public class ProductController {
     @GetMapping("/{productId}")
     public ResponseEntity<ProductDto> getProductById(@PathVariable Long productId) {
         ProductDto productDto = productservice.getProductById(productId);
-//        return new ResponseEntity<ProductDto>(productDto, HttpStatus.OK);
         return ResponseEntity.ok(productDto);
     }
 
@@ -37,10 +36,46 @@ public class ProductController {
         ProductDto savedProduct = productservice.addProduct(productDto);
         return new ResponseEntity<>(savedProduct, HttpStatus.CREATED);
     }
-    //Search products by serach term
+//    Search products by search term
     @GetMapping("/searchproducts")
     public ResponseEntity<List<ProductDto>> searchProductsByTerm(@RequestParam String searchTerm){
         List<ProductDto> matchingProducts = productservice.findProductsBySearchTerm(searchTerm);
         return ResponseEntity.ok(matchingProducts);
+    }
+
+    @PostMapping("/products/category/{categoryId}")
+    public ResponseEntity<ProductDto> addProductWithCategory(@RequestBody ProductDto productDto, @PathVariable Long categoryId) {
+        System.out.println("Add Product with Category endpoint hit with categoryId: " + categoryId);
+        try {
+            ProductDto savedProduct = productservice.addProductWithCategory(productDto, categoryId);
+            return new ResponseEntity<>(savedProduct, HttpStatus.CREATED);
+        } catch (Exception e) {
+            System.out.println("Error adding product: " + e.getMessage());
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+    }
+
+    @PutMapping("{productId}")
+    public ResponseEntity<ProductDto> updateProduct(@PathVariable Long productId, @RequestBody ProductDto productDto) {
+        System.out.println("Update Product endpoint hit with productId: " + productId);
+        try {
+            ProductDto updatedProduct = productservice.updateProduct(productId, productDto);
+            return ResponseEntity.ok(updatedProduct);
+        } catch (Exception e) {
+            System.out.println("Error updating product: " + e.getMessage());
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+    }
+
+    @DeleteMapping("{productId}")
+    public ResponseEntity<Void> deleteProduct(@PathVariable Long productId) {
+        System.out.println("Delete Product endpoint hit with productId: " + productId);
+        try {
+            productservice.deleteProduct(productId);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        } catch (Exception e) {
+            System.out.println("Error deleting product: " + e.getMessage());
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 }
