+++ ranchsupply-backend/ranchsupply-backend/src/main/java/org/launchcode/ranchsupply/service/CategoryService.java
@@ -1,5 +1,6 @@
 package org.launchcode.ranchsupply.service;
 
+import org.launchcode.ranchsupply.exception.ResourceNotFoundException;
 import org.launchcode.ranchsupply.model.Category;
 import org.launchcode.ranchsupply.model.dto.CategoryDto;
 import org.launchcode.ranchsupply.repository.CategoryRepository;
@@ -25,10 +26,30 @@ public class CategoryService {
         return modelMapper.map(savedCategory,CategoryDto.class);
     }
 
+    public CategoryDto getCategoryById(Long categoryId) {
+        Category category = categoryRepository.findById(categoryId)
+                .orElseThrow(() -> new ResourceNotFoundException("Category not found with ID: " + categoryId));
+        return modelMapper.map(category, CategoryDto.class);
+    }
+
     // Get All the categories
     public List<CategoryDto> getAllCategories(){
         List<Category> categories = categoryRepository.findAll();
         return categories.stream().map((category) -> modelMapper.map(category,CategoryDto.class))
                 .collect(Collectors.toList());
+    }
+
+    public CategoryDto updateCategory(Long categoryId, CategoryDto categoryDto) {
+        Category existingCategory = categoryRepository.findById(categoryId)
+                .orElseThrow(() -> new ResourceNotFoundException("Category not found"));
+        modelMapper.map(categoryDto, existingCategory);
+        Category updatedCategory = categoryRepository.save(existingCategory);
+        return modelMapper.map(updatedCategory, CategoryDto.class);
+    }
+
+    public void deleteCategory(Long categoryId) {
+        Category category = categoryRepository.findById(categoryId)
+                .orElseThrow(() -> new ResourceNotFoundException("Category not found"));
+        categoryRepository.delete(category);
     }
 }
