+++ ranchsupply-backend/ranchsupply-backend/src/main/java/org/launchcode/ranchsupply/service/ProductService.java
@@ -1,8 +1,10 @@
 package org.launchcode.ranchsupply.service;
 
 import org.launchcode.ranchsupply.exception.ResourceNotFoundException;
+import org.launchcode.ranchsupply.model.Category;
 import org.launchcode.ranchsupply.model.Product;
 import org.launchcode.ranchsupply.model.dto.ProductDto;
+import org.launchcode.ranchsupply.repository.CategoryRepository;
 import org.launchcode.ranchsupply.repository.ProductRepository;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,6 +17,9 @@ import java.util.stream.Collectors;
 public class ProductService {
     @Autowired
     private ProductRepository productRepository;
+
+    @Autowired
+    private CategoryRepository categoryRepository;
 
     @Autowired
     private ModelMapper modelMapper;
@@ -33,6 +38,15 @@ public class ProductService {
                 .collect(Collectors.toList());
     }
 
+    public ProductDto addProductWithCategory(ProductDto productDto, Long categoryId) {
+        Category category = categoryRepository.findById(categoryId)
+                .orElseThrow(() -> new ResourceNotFoundException("No Category Found"));
+        Product product = modelMapper.map(productDto, Product.class);
+        product.setCategory(category);
+        Product newProduct = productRepository.save(product);
+        return modelMapper.map(newProduct, ProductDto.class);
+    }
+
     // Add a new product
     public ProductDto addProduct(ProductDto productDto) {
         Product product = modelMapper.map(productDto, Product.class);
@@ -44,7 +58,24 @@ public class ProductService {
     public ProductDto updateProduct(Long productId, ProductDto productDto) {
         Product existingProduct = productRepository.findById(productId)
                 .orElseThrow(() -> new ResourceNotFoundException("Product not found"));
-        modelMapper.map(productDto, existingProduct);
+
+        // Update product fields
+        existingProduct.setTitle(productDto.getTitle());
+        existingProduct.setBrand(productDto.getBrand());
+        existingProduct.setDescription(productDto.getDescription());
+        existingProduct.setPrice(productDto.getPrice());
+        existingProduct.setDiscountedPrice(productDto.getDiscountedPrice());
+        existingProduct.setQuantity(productDto.getQuantity());
+        existingProduct.setProductImage(productDto.getProductImage());
+        existingProduct.setSize(productDto.getSize());
+
+        // Update category if provided
+        if (productDto.getCategory() != null && productDto.getCategory().getCategoryId() != null) {
+            Category category = categoryRepository.findById(productDto.getCategory().getCategoryId())
+                    .orElseThrow(() -> new ResourceNotFoundException("Category not found"));
+            existingProduct.setCategory(category);
+        }
+
         Product updatedProduct = productRepository.save(existingProduct);
         return modelMapper.map(updatedProduct, ProductDto.class);
     }
